Work on adding memoization and skeleton/puppet abililty to the S2S compiler.
I would generate the following command (with paths removed) clang++ -Wl,-rpath,install/hclib/lib -shared -undefined dynamic_lookup  -Wl,--no-as-needed -Linstall/hclib/lib -fPIC -Wall -Wno-unused-variable -Wno-unused-parameter test.o -o a.out

and get this error: clang-8: error: no such file or directory: ' -Wl,--no-as-needed'

Then I would copy paste the command into a shell including fish, bash,
and sh and try running it and it worked fine.  I did my best to check
that the space was not a unicode space and I also was able to fix this
issue if I changed the configure.ac from
LDFLAGS = "$LDFLAGS -Wl,--no-as-needed" to
LDFLAGS += "-Wl,--no-as-needed", but that didn't seem great.

anyways it appears that removing the white space solves the problems
whew

Added printing of diff for failed tests when make check is run.

Working towards allowing myself some c++17

Currently the c++17 libraries compile, but they don't get correctly
linked into sstmac_clang so I'll need to work on that.

I don't like this solution, but we can clean it up later.

I couldn't find a nice way to compile specific object files with
different flags. The noinst_* thing worked okay and then you could use
sstmac_clang_LDADD=libclang17.a, but the linker would choose to not
include any of the symbols since non of them were used in sstmac_clang.

This was the fastest way to fix things so that I can continue to work,
with the knowledge that we need to comeback and find a better solution.
I am hopeful that someone with better autotools knowledge (@jjwilke) can
help to find a better solution once the bandwidth is available.

Made a new Pragma that matches the memoize keyword.

It's not as complete as the old one, but I think the idea is cleaner.
Also it's a start. The next step is to get the definiton function added
to the c++ globals and then make sure that any functions that it needs
to call from SST are included/linked.

Print out functions that will be written by the memoization pass.

For now the only thing that you can do is to print the functions.

Memoization work is slowly advancing.

Making progress on memoization

Progress on the memoization expression matcher.